requirements.txt
# Computrabajo Bot — Requirements
# Install with: pip install -r requirements.txt

# CV generation
python-docx>=1.1.0

# Browser automation
playwright>=1.42.0

# AI responses
google-generativeai>=0.5.0

# Environment variables
python-dotenv>=1.0.0
